+++ import-service/src/functions/importFileParser/handler.test.ts
@@ -2,6 +2,8 @@ import { Context } from "aws-lambda";
 import { statusCodesMap, STATUS_MESSAGES } from "src/constants";
 import { importFileParser } from "./handler";
 
+//TODO: fix
+
 const PARAMS = {
   event: {},
   context: {} as Context,
