+++ import-service/src/functions/importProductsFile/mock.json
@@ -5,5 +5,8 @@
     "Access-Control-Allow-Headers": "*",
     "Access-Control-Allow-Origin": "*"
   },
-  "body": "{\"message\":\"SUCCESS\",\"data\":{\"result\":\"https://csv-store-1.s3.eu-west-1.amazonaws.com/uploaded/superfile.csv?AWSAccessKeyId=AKIARP5C2HCUTA5FXTUD&Content-Type=text%2Fcsv&Expires=1620059807&Signature=HnC0FzgIK4KLaW%2FepAQ%2B7ECMzd4%3D\"}}"
+  "body": {
+    "message": "SUCCESS",
+    "data": "https://csv-store-1.s3.eu-west-1.amazonaws.com/uploaded/superfile.csv?AWSAccessKeyId=AKIARP5C2HCUTA5FXTUD&Content-Type=text%2Fcsv&Expires=1620059807&Signature=HnC0FzgIK4KLaW%2FepAQ%2B7ECMzd4%3D"
+  }
 }
