--- import-service/src/schemas/createProductSchema.json
@@ -1,38 +1,0 @@
-{
-  "title": "CreateProductValidation",
-  "description": "Payload used to validate create product functionality",
-  "type": "object",
-  "properties": {
-    "count": {
-      "type": "number"
-    },
-    "description": {
-      "type": "string"
-    },
-    "date": {
-      "type": "string"
-    },
-    "location": {
-      "type": "string"
-    },
-    "price": {
-      "type": "number"
-    },
-    "title": {
-      "type": "string"
-    },
-    "image": {
-      "type": "string"
-    }
-  },
-  "required": [
-    "count",
-    "description",
-    "price",
-    "date",
-    "location",
-    "price",
-    "title",
-    "image"
-  ]
-}
