--- import-service/static-example/index.html
@@ -1,59 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <meta http-equiv="X-UA-Compatible" content="IE=edge">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>Photos</title>
-  <style>
-    h1 {
-      font-size: 18px;
-      font-weight: 800;
-      margin-bottom: 10px;
-    }
-
-    .container {
-      display: flex;
-      flex-direction: row;
-      flex-wrap: wrap;
-      justify-content: center;
-      align-items: center;
-      align-content: center;
-    }
-
-    .imageBox {
-      width: 200px;
-      height: 200px;
-      margin-right: 5px;
-      margin-bottom: 5px;
-      border: 3px solid lightblue;
-    }
-
-    .image {
-      width: 100%;
-      display: block;
-    }
-  </style>
-</head>
-<body>
-  <h1>Photos</h1>
-  <div class="container"></div>
-  <script>
-    const ENDPOINT = "https://k3enib1dug.execute-api.eu-west-1.amazonaws.com/dev/photo-list";
-
-    fetch(ENDPOINT)
-      .then((response) => response.json())
-      .then((images) => {
-        const content = images.data.map(image => {
-          return (`
-            <div class="imageBox">
-              <img class="image" src="${image}" />
-            </div>
-          `);
-        }).join("");
-
-        document.querySelector(".container").innerHTML = content;
-      })
-  </script>
-</body>
-</html>
